Add event_cards
Add teleportation all players by using event card

Add forcing other cards to be sold if they are on loan and not in monopoly

Add forcing auction for card. In this case money for card and money for event_card is returned to user. If card is on loan, only loan_amount returned.